requirements.txt
streamlit
langchain
langchain-chroma
pypdf
python-dotenv
openai
langchain_core